# Conflicts:
#   frontend/src/app/components/recipe/recipe-detail/recipe-detail.component.html

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/validators/RecipeValidator.java
@@ -4,9 +4,6 @@ import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.RecipeCategoryDto;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.RecipeCreateDto;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.RecipeIngredientDto;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.RecipeStepDto;
-import at.ac.tuwien.sepr.groupphase.backend.endpoint.dto.UserRegisterDto;
-import at.ac.tuwien.sepr.groupphase.backend.entity.Category;
-import at.ac.tuwien.sepr.groupphase.backend.entity.RecipeIngredient;
 import at.ac.tuwien.sepr.groupphase.backend.exception.ValidationException;
 import at.ac.tuwien.sepr.groupphase.backend.repository.CategoryRepository;
 import at.ac.tuwien.sepr.groupphase.backend.repository.IngredientRepository;
@@ -36,12 +33,12 @@ public class RecipeValidator {
         List<String> validationErrors = new ArrayList<>();
 
         for (RecipeCategoryDto category : recipe.getCategories()) {
-            if (!categoryRepository.existsById((long) category.getId())) {
+            if (!categoryRepository.existsById(category.getId())) {
                 validationErrors.add("Category " + category.getId() + " not found");
             }
         }
         for (RecipeIngredientDto ingredient : recipe.getIngredients()) {
-            if (!ingredientRepository.existsById((long) ingredient.getId())) {
+            if (!ingredientRepository.existsById(ingredient.getId())) {
                 validationErrors.add("Ingredient " + ingredient.getId() + " not found");
             }
         }
